Revert "Mixed TLS with non-TLS connections."
This reverts commit 8a505fbb01508ce7baac42b0af16fcd40437b8b4.
