+++ go/config/config.go
@@ -88,8 +88,6 @@ type Configuration struct {
 	MySQLTopologySSLCAFile                     string // Certificate Authority PEM file used to authenticate with a Topology mysql instance with TLS
 	MySQLTopologySSLSkipVerify                 bool   // If true, do not strictly validate mutual TLS certs for Topology mysql instances
 	MySQLTopologyUseMutualTLS                  bool   // Turn on TLS authentication with the Topology MySQL instances
-	MySQLTopologyUseMixedTLS                   bool   // Mixed TLS and non-TLS authentication with the Topology MySQL instances
-	TLSCacheTTLFactor                          uint   // Factor of InstancePollSeconds that we set as TLS info cache expiracy
 	BackendDB                                  string // EXPERIMENTAL: type of backend db; either "mysql" or "sqlite3"
 	SQLite3DataFile                            string // when BackendDB == "sqlite3", full path to sqlite3 datafile
 	SkipOrchestratorDatabaseUpdate             bool   // When true, do not check backend database schema nor attempt to update it. Useful when you may be running multiple versions of orchestrator, and you only wish certain boxes to dictate the db structure (or else any time a different orchestrator version runs it will rebuild database schema)
@@ -260,7 +258,6 @@ func newConfiguration() *Configuration {
 		MySQLOrchestratorMaxPoolConnections:        128, // limit concurrent conns to backend DB
 		MySQLOrchestratorPort:                      3306,
 		MySQLTopologyUseMutualTLS:                  false,
-		MySQLTopologyUseMixedTLS:                   false,
 		MySQLOrchestratorUseMutualTLS:              false,
 		MySQLConnectTimeoutSeconds:                 2,
 		MySQLOrchestratorReadTimeoutSeconds:        30,
@@ -268,7 +265,6 @@ func newConfiguration() *Configuration {
 		MySQLTopologyReadTimeoutSeconds:            600,
 		MySQLInterpolateParams:                     false,
 		DefaultInstancePort:                        3306,
-		TLSCacheTTLFactor:                          100,
 		InstancePollSeconds:                        5,
 		InstanceWriteBufferSize:                    100,
 		BufferInstanceWrites:                       false,
